+++ apheleia.el
@@ -32,11 +32,14 @@
   :link '(emacs-commentary-link :tag "Commentary" "apheleia"))
 
 (defcustom apheleia-formatters
-  '((beautysh . ("beautysh"
+  '((astyle . ("astyle" (apheleia-formatters-locate-file
+                         "--options" ".astylerc")))
+    (asmfmt . ("asmfmt"))
+    (bean-format . ("bean-format"))
+    (beautysh . ("beautysh"
                  (apheleia-formatters-indent
                   "--tab" "--indent-size" 'sh-basic-offset)
                  "-"))
-    (bean-format . ("bean-format"))
     (black . ("black"
               (when (apheleia-formatters-extension-p "pyi") "--pyi")
               (apheleia-formatters-fill-column "--line-length")
@@ -52,11 +55,14 @@
     (dart-format . ("dart" "format"))
     (elm-format . ("elm-format" "--yes" "--stdin"))
     (fish-indent . ("fish_indent"))
+    (gawk . ("gawk" "-f" "-" "--pretty-print=-"))
     (gofmt . ("gofmt"))
     (gofumpt . ("gofumpt"))
     (goimports . ("goimports"))
     (google-java-format . ("google-java-format" "-"))
     (isort . ("isort" "-"))
+    (jq "jq" "."
+        (apheleia-formatters-js-indent "--tab" "--indent"))
     (lisp-indent . apheleia-indent-lisp-buffer)
     (ktlint . ("ktlint" "--log-level=none" "--stdin" "-F" "-"))
     (latexindent . ("latexindent" "--logfile=/dev/null"))
@@ -190,6 +196,8 @@ rather than using this system."
     (json-mode . prettier-json)
     (json-ts-mode . prettier-json)
     ;; rest are alphabetical
+    (asm-mode . asmfmt)
+    (awk-mode . gawk)
     (bash-ts-mode . shfmt)
     (beancount-mode . bean-format)
     (c++-ts-mode . clang-format)
@@ -224,6 +232,7 @@ rather than using this system."
     (LaTeX-mode . latexindent)
     (lua-mode . stylua)
     (lisp-mode . lisp-indent)
+    (nasm-mode . asmfmt)
     (nix-mode . nixfmt)
     (purescript-mode . purs-tidy)
     (python-mode . black)